server: push method handler locking from server to handlers
This makes it possible to run multiple instances of the same method
handler concurrently, up to the point where it needs to do internal
synchronization. For method handlers that do not modify internal or
global state, this means they can be run concurrently without any
limitations.

In both cases, this should be way better for server throughput and
latency, at the expense of making method handlers handle
synchronization of their internal state themselves (if any).