server_axum: Move addr to serve fn, add listener
This moves the addr out of the arguments for `from_route` and into
`serve`.
It also adds a new fn to start the server: `serve_listener` which allows
starting the server with an existing `TcpListener`.

`serve` now creates the `TcpListener` (and panics in the same way `bind`
would've panic'ed) and calls `serve_listener`.