dxr: implement ser/de helper functions on top of quick-xml v0.30
This results in consistent behaviour when serializing to XML (i.e.
no self-closing tags in the output) in an effort to be compatible
with the widest possible range of XML-RPC implementations.

Additionally, the dxr_client and dxr_server crates now no longer
depend on quick_xml directly, but only use on the new helper functions
from the dxr crate.